+++ CHANGELOG.md
@@ -11,6 +11,7 @@ API and may have breaking changes during a teeny version change.
 ### Added
 - `VM::error_pop` to get the Ruby Exception and remove it from interfering
   with the current thread
+- `VM::exit` to exit the Ruby VM with status code given
 
 ## [0.6.1] - 2019-06-18
 ### Added